// frontend/lib/api/model/exerciseResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Workout API
 * A workout tracking API built with FastAPI
 * OpenAPI spec version: 1.0.0
 */
import type { ExerciseResponseCreatedByUserId } from "./exerciseResponseCreatedByUserId";
import type { ExerciseModality } from "./exerciseModality";
import type { ExerciseResponsePictureUrl } from "./exerciseResponsePictureUrl";
import type { ExerciseResponseUpdatedByUserId } from "./exerciseResponseUpdatedByUserId";

/**
 * Schema for exercise responses.
 */
export interface ExerciseResponse {
  body_part: string;
  created_at: string;
  created_by_user_id?: ExerciseResponseCreatedByUserId;
  id: number;
  is_user_created: boolean;
  modality: ExerciseModality;
  name: string;
  picture_url?: ExerciseResponsePictureUrl;
  updated_at: string;
  updated_by_user_id?: ExerciseResponseUpdatedByUserId;
}